compile warn suppression

+++ src/eyeline/protogen/server/ProtocolSocketBase.hpp
@@ -34,7 +34,7 @@ public:
     {
       char buf[32];
       std::string rv;
-      sprintf(buf,"[%lu]:",dataSize);
+      sprintf(buf,"[%lu]:",static_cast<unsigned long>(dataSize));
       rv=buf;
       for(size_t i=0;i<dataSize;i++)
       {
